mailgenie-ai/requirements.txt
streamlit
openai
language-tool-python
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
langchain
langchain-community
langchain-core
faiss-cpu
tiktoken
setence-transformers